Move comment_client into django_comment_common
