+++ common/djangoapps/django_comment_common/comment_client/comment.py
@@ -1,4 +1,4 @@
-from lms.lib.comment_client import models, settings
+from django_comment_common.comment_client import models, settings
 
 from .thread import Thread, _url_for_flag_abuse_thread, _url_for_unflag_abuse_thread
 from .utils import CommentClientRequestError, perform_request
