+++ lms/djangoapps/discussion/tasks.py
@@ -11,14 +11,14 @@ from django.contrib.auth.models import User
 from django.contrib.sites.models import Site
 
 from celery_utils.logged_task import LoggedTask
+import django_comment_common.comment_client as cc
 from django_comment_common.models import DiscussionsIdMapping
 from edx_ace import ace
 from edx_ace.utils import date
 from edx_ace.recipient import Recipient
 from eventtracking import tracker
-from opaque_keys.edx.keys import CourseKey
 from lms.djangoapps.django_comment_client.utils import permalink, get_accessible_discussion_xblocks_by_course_id
-import lms.lib.comment_client as cc
+from opaque_keys.edx.keys import CourseKey
 
 from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
 from openedx.core.djangoapps.ace_common.template_context import get_base_template_context
