+++ lms/djangoapps/discussion_api/views.py
@@ -18,11 +18,11 @@ from rest_framework.viewsets import ViewSet
 from six import text_type
 
 from django_comment_client.utils import available_division_schemes
+from django_comment_common import comment_client
 from django_comment_common.models import Role
 from django_comment_common.utils import get_course_discussion_settings, set_course_discussion_settings
 
 from instructor.access import update_forum_role
-from lms.lib import comment_client
 from discussion.views import get_divided_discussions
 from discussion_api.api import (
     create_comment,
