+++ lms/djangoapps/django_comment_client/management/commands/reload_forum_users.py
@@ -3,7 +3,7 @@ Reload forum (comment client) users from existing users.
 """
 from __future__ import print_function
 
-import lms.lib.comment_client as cc
+import django_comment_common.comment_client as cc
 from django.contrib.auth.models import User
 from django.core.management.base import BaseCommand
