+++ lms/djangoapps/django_comment_client/middleware.py
@@ -4,7 +4,7 @@ import logging
 from six import text_type
 
 from django_comment_client.utils import JsonError
-from lms.lib.comment_client import CommentClientRequestError
+from django_comment_common.comment_client import CommentClientRequestError
 
 log = logging.getLogger(__name__)
