+++ lms/djangoapps/django_comment_client/permissions.py
@@ -8,10 +8,10 @@ from types import NoneType
 from edx_django_utils.cache import DEFAULT_REQUEST_CACHE
 from opaque_keys.edx.keys import CourseKey
 
+from django_comment_common.comment_client import Thread
 from django_comment_common.models import CourseDiscussionSettings, all_permissions_for_user_in_course
 from django_comment_common.utils import get_course_discussion_settings
 from lms.djangoapps.teams.models import CourseTeam
-from lms.lib.comment_client import Thread
 from openedx.core.lib.cache_utils import request_cached
