+++ lms/djangoapps/django_comment_client/tests/test_middleware.py
@@ -5,7 +5,7 @@ from django.test import TestCase
 from six import text_type
 
 import django_comment_client.middleware as middleware
-import lms.lib.comment_client
+import django_comment_common.comment_client
 
 
 class AjaxExceptionTestCase(TestCase):
@@ -15,9 +15,9 @@ class AjaxExceptionTestCase(TestCase):
         self.a = middleware.AjaxExceptionMiddleware()
         self.request1 = django.http.HttpRequest()
         self.request0 = django.http.HttpRequest()
-        self.exception1 = lms.lib.comment_client.CommentClientRequestError('{}', 401)
-        self.exception2 = lms.lib.comment_client.CommentClientRequestError('Foo!', 404)
-        self.exception0 = lms.lib.comment_client.CommentClient500Error("Holy crap the server broke!")
+        self.exception1 = django_comment_common.comment_client.CommentClientRequestError('{}', 401)
+        self.exception2 = django_comment_common.comment_client.CommentClientRequestError('Foo!', 404)
+        self.exception0 = django_comment_common.comment_client.CommentClient500Error("Holy crap the server broke!")
         self.request1.META['HTTP_X_REQUESTED_WITH'] = "XMLHttpRequest"
         self.request0.META['HTTP_X_REQUESTED_WITH'] = "SHADOWFAX"
