+++ openedx/core/djangoapps/user_authn/views/tests/test_register.py
@@ -768,8 +768,11 @@ class TestCreateAccountValidation(TestCase):
 
 
 @mock.patch.dict("student.models.settings.FEATURES", {"ENABLE_DISCUSSION_SERVICE": True})
-@mock.patch("lms.lib.comment_client.User.base_url", TEST_CS_URL)
-@mock.patch("lms.lib.comment_client.utils.requests.request", return_value=mock.Mock(status_code=200, text='{}'))
+@mock.patch("django_comment_common.comment_client.User.base_url", TEST_CS_URL)
+@mock.patch(
+    "django_comment_common.comment_client.utils.requests.request",
+    return_value=mock.Mock(status_code=200, text='{}')
+)
 class TestCreateCommentsServiceUser(TransactionTestCase):
     """ Tests for creating comments service user. """
